requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
torch
pytorch-lightning
matplotlib
numpy
kornia
dvc
dvc[gdrive]
dvc[gs]
scikit-learn
glob2
pytorch-lightning
hydra-core
albumentations
torchvision
setuptools==59.5.0
wandb